Minor: Rename scheduling edit and apply changes
