+++ src/app/app.module.ts
@@ -11,9 +11,6 @@ import { AppComponent } from './app.component';
 import { HomeComponent } from './home/home.component';
 import { NavbarHomeComponent } from './layouts/navbar/navbar.component';
 import { CarouselComponent } from './layouts/carousel/carousel.component';
-import { ChecklistProductionComponent } from './checklist/checklist-production/checklist-production.component';
-import { ChecklistMenuComponent } from './checklist/checklist-menu/checklist-menu.component';
-import { ChecklistCafeteriaComponent } from './checklist/checklist-cafeteria/checklist-cafeteria.component';
 import { TextMaskModule } from 'angular2-text-mask';
 import { ROUTES } from './app.routes';
 import { BaseRequestOptions } from '@angular/http';
@@ -23,7 +20,6 @@ import { FooterComponent } from './layouts/footer/footer.component';
 import { AlertComponent } from './layouts/alert/alert.component';
 import { AlertService } from './services/alert/alert.service';
 import { SearchCouncilGroupComponent } from './council-group/search/search-council-group.component';
-import { ChecklistService } from './services/index';
 import {ConfirmationPopoverModule} from 'angular-confirmation-popover';
 import { AlertModule } from 'ngx-bootstrap/alert';
 import { SidebarComponent } from './layouts/sidebar/sidebar.component';
