+++ src/app/app.routes.ts
@@ -1,13 +1,13 @@
 import { SidebarComponent } from './layouts/sidebar/sidebar.component';
-import { EditComponent } from './scheduling/edit/edit.component';
-import { CreateSchedulingComponent } from './scheduling/create/create-scheduling.component';
+import { SchedulingEditComponent } from './scheduling/edit/scheduling-edit.component';
+import { SchedulingCreateComponent } from './scheduling/create/scheduling-create.component';
 import { Routes } from '@angular/router';
 import { ChecklistMenuComponent } from './checklist/checklist-menu/checklist-menu.component';
 import { HomeComponent } from './home/home.component';
 import { ChecklistProductionComponent } from './checklist/checklist-production/checklist-production.component';
 import { ChecklistCafeteriaComponent } from './checklist/checklist-cafeteria/checklist-cafeteria.component';
 import { ChecklistComponent } from './checklist/checklist.component';
-import { SchedulingHomeComponent } from './scheduling/scheduling-home/scheduling-home.component';
+import { SchedulingHomeComponent } from './scheduling/home/scheduling-home.component';
 import { UserListComponent } from './user/list/user-list.component';
 import { CreateUserComponent } from './user/create/create-user.component';
 import { UserEditComponent } from './user/edit/user-edit.component';
@@ -24,8 +24,8 @@ export const ROUTES: Routes = [
   { path: 'checklist/cantina', component: ChecklistCafeteriaComponent},
   { path: 'login', component: LoginComponent },
   { path: 'agendamento', component: SchedulingHomeComponent },
-  { path: 'agendamento/criar', component: CreateSchedulingComponent},
-  { path: 'agendamento/editar/:id', component: EditComponent },
+  { path: 'agendamento/criar', component: SchedulingCreateComponent},
+  { path: 'agendamento/editar/:id', component: SchedulingEditComponent },
   { path: 'usuarios', component: UserListComponent},
   { path: 'usuarios/todos', component: UserListComponent},
   { path: 'login', component: LoginComponent },
