+++ src/app/scheduling/scheduling.module.ts
@@ -1,15 +1,13 @@
-import { CreateSchedulingComponent } from './create/create-scheduling.component';
-
 import { Router, RouterModule } from '@angular/router';
 import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { SchedulingService } from './../services/scheduling.service';
-import { SchedulingHomeComponent } from './scheduling-home/scheduling-home.component';
-import { EditComponent } from './edit/edit.component';
+import { SchedulingHomeComponent } from './home/scheduling-home.component';
+import { SchedulingEditComponent } from './edit/scheduling-edit.component';
+import { SchedulingCreateComponent } from './create/scheduling-create.component';
 import { SchoolService } from '../services/index';
-
 
 @NgModule({
   imports: [
@@ -19,8 +17,8 @@ import { SchoolService } from '../services/index';
   ],
   declarations: [
     SchedulingHomeComponent,
-    EditComponent,
-    CreateSchedulingComponent
+    SchedulingEditComponent,
+    SchedulingCreateComponent
   ],
   providers: [
     SchedulingService,
